cleaned up file structure
AnimationReplacements is no longer the MEGA file
