+++ CustomEmotesAPI/Core/BoneMapper.cs
@@ -1,401 +1,16 @@
-﻿using System;
-using UnityEngine;
-using System.Collections.Generic;
-using UnityEngine.Animations;
-using EmotesAPI;
-using System.Security;
-using System.Security.Permissions;
-using UnityEngine.Networking;
-using System.Collections;
-using UnityEngine.UIElements;
-using System.Net.NetworkInformation;
-using System.Text;
+﻿using EmotesAPI;
 using GameNetcodeStuff;
-using System.IO;
-using MonoMod.RuntimeDetour;
-using UnityEngine.SceneManagement;
-using Unity.Netcode;
 using LethalEmotesAPI;
-using UnityEngine.Audio;
-using System.Globalization;
-using LethalEmotesApi.Ui;
 using LethalEmotesAPI.Utils;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Text;
 using TMPro;
-using BepInEx.Bootstrap;
-using System.Reflection.Emit;
+using Unity.Netcode;
+using UnityEngine;
+using UnityEngine.Animations;
 
-[assembly: SecurityPermission(SecurityAction.RequestMinimum, SkipVerification = true)]
-internal static class AnimationReplacements
-{
-    internal static GameObject g;
-    internal static void RunAll()
-    {
-        ChangeAnims();
-    }
-    internal static bool setup = false;
-    internal static void Import(GameObject prefab, string skeleton, int[] pos, bool hidemesh = true)
-    {
-        Assets.Load<GameObject>(skeleton).GetComponent<Animator>().runtimeAnimatorController = GameObject.Instantiate<GameObject>(Assets.Load<GameObject>("@CustomEmotesAPI_customemotespackage:assets/animationreplacements/commando.prefab")).GetComponent<Animator>().runtimeAnimatorController;
-        AnimationReplacements.ApplyAnimationStuff(prefab, GameObject.Instantiate(Assets.Load<GameObject>(skeleton)), pos, hidemesh, revertBonePositions: true);
-    }
-    public static void DebugBones(GameObject fab)
-    {
-        var meshes = fab.GetComponentsInChildren<SkinnedMeshRenderer>();
-        StringBuilder sb = new StringBuilder();
-        sb.Append($"rendererererer: {meshes[0]}\n");
-        sb.Append($"bone count: {meshes[0].bones.Length}\n");
-        sb.Append($"mesh count: {meshes.Length}\n");
-        sb.Append($"root bone: {meshes[0].rootBone.name}\n");
-        sb.Append($"{fab.ToString()}:\n");
-        if (meshes[0].bones.Length == 0)
-        {
-            sb.Append("No bones");
-        }
-        else
-        {
-            sb.Append("[");
-            foreach (var bone in meshes[0].bones)
-            {
-                sb.Append($"'{bone.name}', ");
-            }
-            sb.Remove(sb.Length - 2, 2);
-            sb.Append("]");
-        }
-        sb.Append("\n\n");
-        DebugClass.Log(sb.ToString());
-    }
-    internal static void ChangeAnims()
-    {
-
-    }
-    internal static void ApplyAnimationStuff(GameObject bodyPrefab, string resource, int[] pos)
-    {
-        GameObject animcontroller = Assets.Load<GameObject>(resource);
-        ApplyAnimationStuff(bodyPrefab, animcontroller, pos);
-    }
-
-    internal static void ApplyAnimationStuff(GameObject bodyPrefab, GameObject animcontroller, int[] pos, bool hidemeshes = true, bool jank = false, bool revertBonePositions = false)
-    {
-        try
-        {
-            if (!animcontroller.GetComponentInChildren<Animator>().avatar.isHuman)
-            {
-                DebugClass.Log($"{animcontroller}'s avatar isn't humanoid, please fix it in unity!");
-                return;
-            }
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had issue checking if avatar was humanoid: {e}");
-            throw;
-        }
-        try
-        {
-            if (hidemeshes)
-            {
-                foreach (var item in animcontroller.GetComponentsInChildren<SkinnedMeshRenderer>())
-                {
-                    item.sharedMesh = null;
-                }
-                foreach (var item in animcontroller.GetComponentsInChildren<MeshFilter>())
-                {
-                    item.sharedMesh = null;
-                }
-            }
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had trouble while hiding meshes: {e}");
-            throw;
-        }
-
-        Transform modelTransform;
-        modelTransform = bodyPrefab.GetComponentInChildren<Animator>().transform;
-        try
-        {
-            animcontroller.transform.parent = modelTransform;
-            animcontroller.transform.localPosition = Vector3.zero;
-            animcontroller.transform.localEulerAngles = new Vector3(90, 0, 0);
-            animcontroller.transform.localScale = Vector3.one;
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had trouble setting emote skeletons parent: {e}");
-            throw;
-        }
-
-        SkinnedMeshRenderer smr1;
-        SkinnedMeshRenderer[] smr2 = new SkinnedMeshRenderer[pos.Length];
-        try
-        {
-            smr1 = animcontroller.GetComponentsInChildren<SkinnedMeshRenderer>()[0];
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had trouble setting emote skeletons SkinnedMeshRenderer: {e}");
-            throw;
-        }
-        try
-        {
-            for (int i = 0; i < pos.Length; i++)
-            {
-                smr2[i] = bodyPrefab.GetComponentsInChildren<SkinnedMeshRenderer>()[pos[i]];
-            }
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had trouble setting the original skeleton's skinned mesh renderer: {e}");
-            throw;
-        }
-
-
-        //since this game is jank and has A UNIQUE SKINNEDMESHRENDERER FOR EACH LOD, I am just going to enforce proper SMR labeling. This probably won't be that big of a deal since I imagine the need for people setting up their own emote skeletons will be FAR less than ROR2
-        //try
-        //{
-        //    int matchingBones = 0;
-        //    while (true)
-        //    {
-        //        foreach (var smr1bone in smr1.bones) //smr is SkinnedMeshRenderer
-        //        {
-        //            foreach (var smr2bone in smr2.bones)
-        //            {
-        //                if (smr1bone.name == smr2bone.name)
-        //                {
-        //                    matchingBones++;
-        //                }
-        //            }
-        //        }
-        //        if (matchingBones < 1 && pos + 1 < bodyPrefab.GetComponentsInChildren<SkinnedMeshRenderer>().Length)
-        //        {
-        //            pos++;
-        //            smr2 = bodyPrefab.GetComponentsInChildren<SkinnedMeshRenderer>()[pos];
-        //            matchingBones = 0;
-        //        }
-        //        else
-        //        {
-        //            break;
-        //        }
-        //    }
-        //}
-        //catch (Exception e)
-        //{
-        //    DebugClass.Log($"Had issue while checking matching bones: {e}");
-        //    throw;
-        //}
-
-        var test = animcontroller.AddComponent<BoneMapper>();
-        try
-        {
-            test.jank = jank;
-            test.emoteSkeletonSMR = smr1;
-            test.basePlayerModelSMR = smr2;
-            test.bodyPrefab = bodyPrefab;
-            test.basePlayerModelAnimator = modelTransform.GetComponentInChildren<Animator>();
-            test.emoteSkeletonAnimator = animcontroller.GetComponentInChildren<Animator>();
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had issue when setting up BoneMapper settings 1: {e}");
-            throw;
-        }
-        try
-        {
-            var nuts = Assets.Load<GameObject>("assets/customstuff/scavEmoteSkeleton.prefab");
-            float banditScale = Vector3.Distance(nuts.GetComponentInChildren<Animator>().GetBoneTransform(HumanBodyBones.Head).position, nuts.GetComponentInChildren<Animator>().GetBoneTransform(HumanBodyBones.LeftFoot).position);
-            float currScale = Vector3.Distance(animcontroller.GetComponentInChildren<Animator>().GetBoneTransform(HumanBodyBones.Head).position, animcontroller.GetComponentInChildren<Animator>().GetBoneTransform(HumanBodyBones.LeftFoot).position);
-            test.scale = currScale / banditScale;
-            test.h = bodyPrefab.GetComponentInChildren<PlayerControllerB>().health;
-            test.model = modelTransform.gameObject;
-        }
-        catch (Exception e)
-        {
-            DebugClass.Log($"Had issue when setting up BoneMapper settings 2: {e}");
-            throw;
-        }
-        test.revertTransform = revertBonePositions;
-
-    }
-}
-public struct JoinSpot
-{
-    public string name;
-    public Vector3 position;
-    public Vector3 rotation;
-    public Vector3 scale;
-    public JoinSpot(string _name, Vector3 _position, Vector3 _rotation, Vector3 _scale)
-    {
-        name = _name;
-        position = _position;
-        rotation = _rotation;
-        scale = _scale;
-    }
-    public JoinSpot(string _name, Vector3 _position)
-    {
-        name = _name;
-        position = _position;
-        rotation = Vector3.zero;
-        scale = Vector3.one;
-    }
-}
-public class CustomAnimationClip : MonoBehaviour
-{
-    public AnimationClip[] clip, secondaryClip; //DONT SUPPORT MULTI CLIP ANIMATIONS TO SYNC     //but why not? how hard could it be, I'm sure I left that note for a reason....  //it was for a reason, but it works now
-    public bool looping;
-    public string wwiseEvent;
-    public bool syncronizeAudio;
-    public List<HumanBodyBones> soloIgnoredBones;
-    public List<HumanBodyBones> rootIgnoredBones;
-    public bool dimAudioWhenClose;
-    public bool stopOnAttack;
-    public bool stopOnMove;
-    public bool visibility;
-    public int startPref, joinPref;
-    public JoinSpot[] joinSpots;
-    public bool useSafePositionReset;
-    public string customName;
-    public Action<BoneMapper> customPostEventCodeSync;
-    public Action<BoneMapper> customPostEventCodeNoSync;
-
-
-    public bool syncronizeAnimation;
-    public int syncPos;
-    public static List<float> syncTimer = new List<float>();
-    public static List<int> syncPlayerCount = new List<int>();
-    public static List<List<bool>> uniqueAnimations = new List<List<bool>>();
-    public bool vulnerableEmote = false;
-    public AnimationClipParams.LockType lockType = AnimationClipParams.LockType.none;
-    public bool willGetClaimed = false;
-    public float audioLevel = .5f;
-    public bool thirdPerson = false;
-
-    internal CustomAnimationClip(AnimationClip[] _clip, bool _loop, AudioClip[] primaryAudioClips = null, AudioClip[] secondaryAudioClips = null, HumanBodyBones[] rootBonesToIgnore = null, HumanBodyBones[] soloBonesToIgnore = null, AnimationClip[] _secondaryClip = null, bool dimWhenClose = false, bool stopWhenMove = false, bool stopWhenAttack = false, bool visible = true, bool syncAnim = false, bool syncAudio = false, int startPreference = -1, int joinPreference = -1, JoinSpot[] _joinSpots = null, bool safePositionReset = false, string customName = "", Action<BoneMapper> _customPostEventCodeSync = null, Action<BoneMapper> _customPostEventCodeNoSync = null, AnimationClipParams.LockType lockType = AnimationClipParams.LockType.none, AudioClip[] primaryDMCAFreeAudioClips = null, AudioClip[] secondaryDMCAFreeAudioClips = null, bool willGetClaimed = false, float audioLevel = .5f, bool thirdPerson = false)
-    {
-        if (rootBonesToIgnore == null)
-            rootBonesToIgnore = new HumanBodyBones[0];
-        if (soloBonesToIgnore == null)
-            soloBonesToIgnore = new HumanBodyBones[0];
-        clip = _clip;
-        secondaryClip = _secondaryClip;
-        looping = _loop;
-        dimAudioWhenClose = dimWhenClose;
-        stopOnAttack = stopWhenAttack;
-        stopOnMove = stopWhenMove;
-        visibility = visible;
-        joinPref = joinPreference;
-        startPref = startPreference;
-        customPostEventCodeSync = _customPostEventCodeSync;
-        customPostEventCodeNoSync = _customPostEventCodeNoSync;
-        if (primaryAudioClips == null)
-        {
-            BoneMapper.primaryAudioClips.Add(new AudioClip[] { null });
-        }
-        else
-        {
-            BoneMapper.primaryAudioClips.Add(primaryAudioClips);
-        }
-        if (secondaryAudioClips == null)
-        {
-            BoneMapper.secondaryAudioClips.Add(new AudioClip[] { null });
-        }
-        else
-        {
-            BoneMapper.secondaryAudioClips.Add(secondaryAudioClips);
-        }
-
-        if (primaryDMCAFreeAudioClips == null)
-        {
-            BoneMapper.primaryDMCAFreeAudioClips.Add(new AudioClip[] { null });
-        }
-        else
-        {
-            BoneMapper.primaryDMCAFreeAudioClips.Add(primaryDMCAFreeAudioClips);
-        }
-        if (secondaryDMCAFreeAudioClips == null)
-        {
-            BoneMapper.secondaryDMCAFreeAudioClips.Add(new AudioClip[] { null });
-        }
-        else
-        {
-            BoneMapper.secondaryDMCAFreeAudioClips.Add(secondaryDMCAFreeAudioClips);
-        }
-
-        if (soloBonesToIgnore.Length != 0)
-        {
-            soloIgnoredBones = new List<HumanBodyBones>(soloBonesToIgnore);
-        }
-        else
-        {
-            soloIgnoredBones = new List<HumanBodyBones>();
-        }
-
-        if (rootBonesToIgnore.Length != 0)
-        {
-            rootIgnoredBones = new List<HumanBodyBones>(rootBonesToIgnore);
-        }
-        else
-        {
-            rootIgnoredBones = new List<HumanBodyBones>();
-        }
-        syncronizeAnimation = syncAnim;
-        syncronizeAudio = syncAudio;
-        syncPos = syncTimer.Count;
-        syncTimer.Add(0);
-        syncPlayerCount.Add(0);
-        List<bool> bools = new List<bool>();
-        for (int i = 0; i < _clip.Length; i++)
-        {
-            bools.Add(false);
-        }
-        uniqueAnimations.Add(bools);
-
-        if (_joinSpots == null)
-            _joinSpots = new JoinSpot[0];
-        joinSpots = _joinSpots;
-        this.useSafePositionReset = safePositionReset;
-        this.customName = customName;
-        if (customName != "")
-        {
-            BoneMapper.customNamePairs.Add(customName, _clip[0].name);
-        }
-        else
-        {
-            this.customName = clip[0].name;
-        }
-        BoneMapper.listOfCurrentEmoteAudio.Add(new List<AudioSource>());
-        this.lockType = lockType;
-        this.willGetClaimed = willGetClaimed;
-        this.audioLevel = audioLevel;
-        this.thirdPerson = thirdPerson;
-    }
-    private static GameObject audioLoader;
-}
-public struct WorldProp
-{
-    internal GameObject prop;
-    internal JoinSpot[] joinSpots;
-    public WorldProp(GameObject _prop, JoinSpot[] _joinSpots)
-    {
-        prop = _prop;
-        joinSpots = _joinSpots;
-    }
-}
-public enum TempThirdPerson
-{
-    none,
-    on,
-    off
-}
-public class AudioObject : MonoBehaviour
-{
-    internal int spot;
-    internal int playerCount;
-    internal GameObject audioObject;
-    internal int activeObjectsSpot;
-}
-public class AudioContainer : MonoBehaviour
-{
-    internal List<GameObject> playingObjects = new List<GameObject>();
-}
 public class BoneMapper : MonoBehaviour
 {
     public static List<AudioClip[]> primaryAudioClips = new List<AudioClip[]>();
@@ -908,9 +523,6 @@ public class BoneMapper : MonoBehaviour
                         {
                             startingXPoint = x;
                             //DebugClass.Log($"they are equal!");
-                            var s = new ConstraintSource();
-                            s.sourceTransform = emoteSkeletonSMR.bones[i];
-                            s.weight = 1;
                             //DebugClass.Log($"{smr.name}--- i is equal to {x}  ------ {smr.bones[x].name}");
                             EmoteConstraint e = smr.bones[x].gameObject.AddComponent<EmoteConstraint>();
                             e.AddSource(ref smr.bones[x], ref emoteSkeletonSMR.bones[i]);
@@ -1571,19 +1183,5 @@ public class BoneMapper : MonoBehaviour
             mapperBody.grabDistance = 3f;
             isInThirdPerson = false;
         }
-    }
-}
-public class BonePair
-{
-    public Transform original, newiginal;
-    public BonePair(Transform n, Transform o)
-    {
-        newiginal = n;
-        original = o;
-    }
-
-    public void test()
-    {
-
     }
 }
