+++ CustomEmotesAPI/Core/EmoteConstraint.cs
