+++ CustomEmotesAPI/Core/EmoteLocation.cs
@@ -6,6 +6,27 @@ using System.Text;
 using UnityEngine;
 using UnityEngine.Networking;
 
+public struct JoinSpot
+{
+    public string name;
+    public Vector3 position;
+    public Vector3 rotation;
+    public Vector3 scale;
+    public JoinSpot(string _name, Vector3 _position, Vector3 _rotation, Vector3 _scale)
+    {
+        name = _name;
+        position = _position;
+        rotation = _rotation;
+        scale = _scale;
+    }
+    public JoinSpot(string _name, Vector3 _position)
+    {
+        name = _name;
+        position = _position;
+        rotation = Vector3.zero;
+        scale = Vector3.one;
+    }
+}
 public class EmoteLocation : MonoBehaviour
 {
     public static List<EmoteLocation> emoteLocations = new List<EmoteLocation>();
