+++ CustomEmotesAPI/Data/Settings.cs
