+++ CustomEmotesAPI/Unused/AudioContainer.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+//unused in lethal company
+public class AudioContainer : MonoBehaviour
+{
+    internal List<GameObject> playingObjects = new List<GameObject>();
+}
