+++ CustomEmotesAPI/Unused/AudioObject.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+//unused in lethal company
+public class AudioObject : MonoBehaviour
+{
+    internal int spot;
+    internal int playerCount;
+    internal GameObject audioObject;
+    internal int activeObjectsSpot;
+}
