+++ CustomEmotesAPI/Unused/BonePair.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+public class BonePair
+{
+    public Transform original, newiginal;
+    public BonePair(Transform n, Transform o)
+    {
+        newiginal = n;
+        original = o;
+    }
+
+    public void test()
+    {
+
+    }
+}
