+++ CustomEmotesAPI/Utils/DebugClass.cs
