+++ CustomEmotesAPI/Utils/EmotesInputSettings.cs
