+++ CustomEmotesAPI/Utils/Networking.cs
