+++ CustomEmotesAPI/Utils/WorldPropSpawnHandler.cs
